Moves lift.py functionality to a new anvill python module __main__.py
