+++ CMakeLists.txt
@@ -123,6 +123,7 @@ target_link_libraries(${SPECIFY_BITCODE} PRIVATE ${ANVILL})
 set(ANVILL_PYTHON_SOURCES
     setup.py
     python/anvill/__init__.py
+    python/anvill/__main__.py
     python/anvill/arch.py
     python/anvill/binja.py
     python/anvill/dwarf.py
